2ο Εξάμηνο/Αποκεντρωμένες Τεχνολογίες/Assignment 1/final/requirements.txt
absl-py==0.7.0
alabaster==0.7.12
apt-clone==0.2.1
apturl==0.5.2
asn1crypto==0.24.0
astroid==2.1.0
attrs==18.2.0
Babel==2.6.0
backcall==0.1.0
base58check==1.0.2
beautifulsoup4==4.6.0
bitcoin==1.1.42
bitcoin-utils==0.3.1
bleach==1.5.0
boto==2.49.0
boto3==1.9.101
botocore==1.12.101
Brlapi==0.6.6
bz2file==0.98
certifi==2018.1.18
cffi==1.12.1
chardet==3.0.4
Click==7.0
cloudpickle==0.8.0
command-not-found==0.3
configobj==5.0.6
costcla==0.5
cryptography==2.5
cupshelpers==1.0
cycler==0.10.0
decorator==4.3.2
defer==1.0.6
defusedxml==0.5.0
distro-info===0.18ubuntu0.18.04.1
docutils==0.14
ecdsa==0.13
entrypoints==0.3
feedparser==5.2.1
Flask==1.0.2
gensim==3.7.1
gpodder==3.10.1
graphviz==0.10.1
h5py==2.9.0
html5lib==0.9999999
httplib2==0.9.2
idna==2.6
imagesize==1.1.0
imbalanced-learn==0.4.3
imblearn==0.0
ipykernel==5.1.0
ipython==7.3.0
ipython-genutils==0.2.0
ipywidgets==7.4.2
isort==4.3.4
itsdangerous==1.1.0
jedi==0.13.3
jeepney==0.4
Jinja2==2.10
jmespath==0.9.3
jsonrpc-requests==0.4.0
jsonschema==3.0.0
jupyter==1.0.0
jupyter-client==5.2.4
jupyter-console==6.0.0
jupyter-core==4.4.0
jupyterlab==0.35.4
jupyterlab-server==0.2.0
kazam==1.4.5
Keras==2.2.4
Keras-Applications==1.0.7
Keras-Preprocessing==1.0.9
keyring==18.0.0
kiwisolver==1.0.1
lazy-object-proxy==1.3.1
louis==3.5.0
macaroonbakery==1.1.3
Mako==1.0.7
Markdown==3.0.1
MarkupSafe==1.0
matplotlib==3.0.2
mccabe==0.6.1
mistune==0.8.4
mlxtend==0.15.0.0
mpmath==1.1.0
mutagen==1.41.1
mygpoclient==1.8
nbconvert==5.4.1
nbformat==4.4.0
netifaces==0.10.4
nltk==3.4
notebook==5.7.4
numpy==1.16.2
numpydoc==0.8.0
onboard==1.4.1
openshot-qt==2.4.1
packaging==19.0
PAM==0.4.2
pandas==0.24.1
pandocfilters==1.4.2
parso==0.3.4
patsy==0.5.1
pexpect==4.2.1
pickleshare==0.7.5
Pillow==5.1.0
podcastparser==0.6.2
prometheus-client==0.6.0
prompt-toolkit==2.0.9
protobuf==3.6.1
psutil==5.4.2
ptyprocess==0.6.0
pycairo==1.16.2
pycodestyle==2.5.0
pycparser==2.19
pycrypto==2.6.1
pycups==1.9.73
pycurl==7.43.0.1
pyea==0.2
pyflakes==2.1.0
Pygments==2.3.1
pygobject==3.26.1
PyICU==1.9.8
pyinotify==0.9.6
pylint==2.2.2
pymacaroons==0.13.0
PyNaCl==1.1.2
pyparsing==2.3.1
PyQt5==5.9.2
pyRFC3339==1.0
pyrsistent==0.14.11
python-apt==1.6.3+ubuntu1
python-bitcoinrpc==1.0
python-dateutil==2.8.0
python-debian==0.1.32
python-xapp==1.2.0
python-xlib==0.20
pytz==2018.3
pyxdg==0.25
PyYAML==3.12
pyzmq==18.0.0
QtAwesome==0.5.6
qtconsole==4.4.3
QtPy==1.6.0
quodlibet==4.2.0
reportlab==3.4.0
requests==2.18.4
requests-unixsocket==0.1.5
rope==0.12.0
s3transfer==0.2.0
scikit-learn==0.19.0
scipy==1.2.1
screen-resolution-extra==0.0.0
seaborn==0.9.0
SecretStorage==3.1.1
Send2Trash==1.5.0
sessioninstaller==0.0.0
setproctitle==1.1.10
singledispatch==3.4.0.3
sip==4.19.8
six==1.11.0
sklearn==0.0
smart-open==1.8.0
snowballstemmer==1.2.1
Sphinx==1.8.4
sphinxcontrib-websupport==1.1.0
spyder==3.3.2
spyder-kernels==0.4.2
statsmodels==0.9.0
sympy==1.3
system-service==0.3
systemd-python==234
tabulate==0.8.3
tensorflow==1.5.0
tensorflow-tensorboard==1.5.1
terminado==0.8.1
testpath==0.4.2
tornado==5.1.1
traitlets==4.3.2
typed-ast==1.3.1
ubuntu-drivers-common==0.0.0
ufw==0.36
urllib3==1.22
virtualenv==15.1.0
wcwidth==0.1.7
Werkzeug==0.14.1
widgetsnbextension==3.4.2
wrapt==1.11.1
wurlitzer==1.0.2
xkit==0.0.0
